Extract faces.push into faces/push.ts module #5506
