+++ impl/src/main/ts/faces.ts
@@ -1,6 +1,7 @@
 // @ts-nocheck — legacy monolithic file; type-checking enforced module-by-module during the split.
 import { getHead, getNonce, executeScriptWithNonce } from "./faces/dom";
 import { chain as utilChain } from "./faces/util";
+import { push as facesPush } from "./faces/push";
 
 // Detect if this is already loaded, and if loaded, if it's a higher version
 if ( !( (window.faces && window.faces.specversion && window.faces.specversion >= 50000 )
@@ -2719,212 +2720,11 @@ if ( !( (window.faces && window.faces.specversion && window.faces.specversion >=
     };
 
     /**
-     * <p class="changed_added_2_3">
-     * The Push functionality.
-     * </p>
+     * <p class="changed_added_2_3">The Push functionality.</p>
      * @name faces.push
      * @namespace
-     * @exec
      */
-    faces.push = (function(window) {
-
-        // "Constant" fields ----------------------------------------------------------------------------------------------
-
-        const RECONNECT_INTERVAL = 500;
-        const MAX_RECONNECT_ATTEMPTS = 25;
-        const REASON_EXPIRED = "Expired";
-        const REASON_UNKNOWN_CHANNEL = "Unknown channel";
-
-        // Private static fields ------------------------------------------------------------------------------------------
-
-        const sockets = {};
-        const self = {};
-
-        // Private constructor functions ----------------------------------------------------------------------------------
-
-        /**
-         * Creates a reconnecting websocket. When the websocket successfully connects on first attempt, then it will
-         * automatically reconnect on timeout with cumulative intervals of 500ms with a maximum of 25 attempts (~3 minutes).
-         * The <code>onclose</code> function will be called with the error code of the last attempt.
-         * @constructor
-         * @param {string} url The URL of the websocket.
-         * @param {string} channel The channel name of the websocket.
-         * @param {function} onopen The function to be invoked when the websocket is opened.
-         * @param {function} onmessage The function to be invoked when a message is received.
-         * @param {function} onerror The function to be invoked when a connection error has occurred and the web socket will attempt to reconnect.
-         * @param {function} onclose The function to be invoked when the web socket is closed and will not anymore attempt to reconnect.
-         * @param {Object} behaviors Client behavior functions to be invoked when specific message is received.
-         */
-        function ReconnectingWebsocket(url, channel, onopen, onmessage, onerror, onclose, behaviors) {
-
-            // Private fields -----------------------------------------------------------------------------------------
-
-            let socket;
-            let reconnectAttempts;
-            const self = this;
-
-            // Public functions ---------------------------------------------------------------------------------------
-
-            /**
-             * Opens the reconnecting websocket.
-             */
-            self.open = function() {
-                if (socket && socket.readyState === 1) {
-                    return;
-                }
-
-                socket = new WebSocket(url);
-
-                socket.onopen = function(event) {
-                    if (reconnectAttempts == null) {
-                        onopen(channel);
-                    }
-
-                    reconnectAttempts = 0;
-                }
-
-                socket.onmessage = function(event) {
-                    const message = JSON.parse(event.data).data;
-                    onmessage(message, channel, event);
-                    const functions = behaviors[message];
-
-                    if (functions && functions.length) {
-                        for (let i = 0; i < functions.length; i++) {
-                            functions[i]();
-                        }
-                    }
-                }
-
-                socket.onclose = function(event) {
-                    if (!socket
-                        || (event.code === 1000 && event.reason === REASON_EXPIRED)
-                        || (event.code === 1008 || event.reason === REASON_UNKNOWN_CHANNEL) // Older IE versions incorrectly return 1005 instead of 1008, hence the fallback check on the message.
-                        || (reconnectAttempts == null)
-                        || (reconnectAttempts >= MAX_RECONNECT_ATTEMPTS))
-                    {
-                        onclose(event.code, channel, event);
-                    }
-                    else {
-                        onerror(event.code, channel, event);
-                        setTimeout(self.open, RECONNECT_INTERVAL * reconnectAttempts++);
-                    }
-                }
-            }
-
-            /**
-             * Closes the reconnecting websocket.
-             */
-            self.close = function() {
-                if (socket) {
-                    const s = socket;
-                    socket = null;
-                    reconnectAttempts = null;
-                    s.close();
-                }
-            }
-
-        }
-
-        // Public static functions ----------------------------------------------------------------------------------------
-
-        /**
-         * Initialize a websocket on the given client identifier. When connected, it will stay open and reconnect as
-         * long as URL is valid and <code>faces.push.close()</code> hasn't explicitly been called on the same client
-         * identifier.
-         * @param {string} clientId The client identifier of the websocket.
-         * @param {string} url The URL of the websocket. All open websockets on the same URL will receive the
-         * same push notification from the server.
-         * @param {string} channel The channel name of the websocket.
-         * @param {function} onopen The JavaScript event handler function that is invoked when the websocket is opened.
-         * The function will be invoked with one argument: the channel name.
-         * @param {function} onmessage The JavaScript event handler function that is invoked when a message is received from
-         * the server. The function will be invoked with three arguments: the push message, the channel name and
-         * the raw <code>MessageEvent</code> itself.
-         * @param {function} onerror The JavaScript event handler function that is invoked when a connection error has
-         * occurred and the web socket will attempt to reconnect. The function will be invoked with three arguments: the
-         * error reason code, the channel name and the raw <code>CloseEvent</code> itself. Note that this will not be
-         * invoked on final close of the web socket, even when the final close is caused by an error. See also
-         * <a href="http://tools.ietf.org/html/rfc6455#section-7.4.1">RFC 6455 section 7.4.1</a> and {@link CloseCodes} API
-         * for an elaborate list of all close codes.
-         * @param {function} onclose The function to be invoked when the web socket is closed and will not anymore attempt
-         * to reconnect. The function will be invoked with three arguments: the close reason code, the channel name
-         * and the raw <code>CloseEvent</code> itself. Note that this will also be invoked when the close is caused by an
-         * error and that you can inspect the close reason code if an actual connection error occurred and which one (i.e.
-         * when the code is not 1000 or 1008). See also <a href="http://tools.ietf.org/html/rfc6455#section-7.4.1">RFC 6455
-         * section 7.4.1</a> and {@link CloseCodes} API for an elaborate list of all close codes.
-         * @param {Object} behaviors Client behavior functions to be invoked when specific message is received.
-         * @param {boolean} autoconnect Whether or not to automatically connect the socket. Defaults to <code>false</code>.
-         * @member faces.push
-         * @function faces.push.init
-         */
-        self.init = function(clientId, url, channel, onopen, onmessage, onerror, onclose, behaviors, autoconnect) {
-            onclose = resolveFunction(onclose);
-
-            if (!window.WebSocket) {
-                onclose(-1, clientId);
-                return;
-            }
-
-            if (!sockets[clientId]) {
-                sockets[clientId] = new ReconnectingWebsocket(url, channel, resolveFunction(onopen), resolveFunction(onmessage), resolveFunction(onerror), onclose, behaviors);
-            }
-
-            if (autoconnect) {
-                self.open(clientId);
-            }
-        }
-
-        /**
-         * Open the websocket on the given client identifier.
-         * @param {string} clientId The client identifier of the websocket.
-         * @throws {Error} When client identifier is unknown. You may need to initialize it first via <code>init()</code> function.
-         * @member faces.push
-         * @function faces.push.open
-         */
-        self.open = function(clientId) {
-            getSocket(clientId).open();
-        }
-
-        /**
-         * Close the websocket on the given client identifier.
-         * @param {string} clientId The client identifier of the websocket.
-         * @throws {Error} When client identifier is unknown. You may need to initialize it first via <code>init()</code> function.
-         * @member faces.push
-         * @function faces.push.close
-         */
-        self.close = function(clientId) {
-            getSocket(clientId).close();
-        }
-
-        // Private static functions ---------------------------------------------------------------------------------------
-
-        /**
-         * If given function is actually not a function, then try to interpret it as name of a global function.
-         * If it still doesn't resolve to anything, then return a NOOP function.
-         * @param {Object} fn Can be function, or string representing function name, or undefined.
-         * @return {function} The intended function, or a NOOP function when undefined.
-         */
-        function resolveFunction(fn) {
-            return (typeof fn !== "function") && (fn = window[fn] || function(){}), fn;
-        }
-
-        /**
-         * Get socket associated with given client identifier.
-         * @param {string} clientId The client identifier of the websocket.
-         * @return {Socket} Socket associated with given client identifier.
-         * @throws {Error} When client identifier is unknown. You may need to initialize it first via <code>init()</code> function.
-         */
-        function getSocket(clientId) {
-            const socket = sockets[clientId];
-            if (socket) return socket;
-            else throw new Error("Unknown clientId: " + clientId);
-        }
-
-        // Expose self to public ------------------------------------------------------------------------------------------
-
-        return self;
-
-    })(window);
+    faces.push = facesPush;
 
 
     /**
